Split into main and questions pages
The existing site is a single page with divs for the main and questions
"pages" which are hidden and shown as necessary. There are a few issues
with this approach:

  - Currently, elements seem to be hidden using display: none, which has
  accessibility issues, i.e. might not work for people who use screen
  readers:
  https://developer.mozilla.org/en-US/docs/Web/CSS/display#Accessibility_concern

  - Since all the javascript is in a single file, it's tempting to use
  global variables. This makes the code difficult to change, since we
  can't be sure changes to one page wont affect another.

  - You can't link to the questions page directly because the url does
  not change after submit is clicked

  - Editing the questions page requires a lot of scrolling in your text
  editor

There are methods around all of these concerns, and indeed many sites
are built as single page applications. But for now, splitting into
separate pages is simpler.